Improve connection close and rollback scenarios
Once all the updates of a particular database are successfully executed,
database connection is committed and closed. In an error scenario only
the current db connection is rolled back and closed(as other
successfully updated db connections are already closed). In an error
scenario user can re run the tool to update entries in the failed db. In
this scenario, user needs to provide the pseudonym(which is printed only
in the console in the previous run) as an input when re running the
tool.